+++ dundie/cli.py
@@ -1,0 +1,25 @@
+import argparse
+from .core import load
+def main():
+    parse = argparse.ArgumentParser(
+    description="Dunder Mifflin Rewards CLI",
+    epilog="Enjoy and use with cautious.", 
+    )
+    parse.add_argument(
+        "subcommand",
+        type=str,
+        help="The subcommand to run",
+        choices=("load","show","send"),
+        default="help",
+        
+    )
+    parse.add_argument(
+        "filepath",
+        type=str,
+        help="File path to load",
+        default=None
+    )
+    args = parse.parse_args()
+  
+    globals()[args.subcommand](args.filepath)
+
