+++ dundie/utils/log.py
@@ -1,0 +1,6 @@
+import os 
+import logging
+
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.debug)
+logging.FileHandler("meulog.log")
