+++ requirements.test.txt
@@ -1,1 +1,1 @@
-ipython
+pytest
